+++ src/Providers/DeepSeek/Handlers/Text.php
@@ -10,8 +10,8 @@ use EchoLabs\Prism\Providers\DeepSeek\Maps\MessageMap;
 use EchoLabs\Prism\Providers\DeepSeek\Maps\ToolCallMap;
 use EchoLabs\Prism\Providers\DeepSeek\Maps\ToolChoiceMap;
 use EchoLabs\Prism\Providers\DeepSeek\Maps\ToolMap;
-use EchoLabs\Prism\Providers\ProviderResponse;
 use EchoLabs\Prism\Text\Request;
+use EchoLabs\Prism\ValueObjects\ProviderResponse;
 use EchoLabs\Prism\ValueObjects\Usage;
 use Illuminate\Http\Client\PendingRequest;
 use Illuminate\Http\Client\Response;
