+++ src/Providers/Groq/Handlers/Text.php
@@ -9,8 +9,8 @@ use EchoLabs\Prism\Providers\Groq\Maps\FinishReasonMap;
 use EchoLabs\Prism\Providers\Groq\Maps\MessageMap;
 use EchoLabs\Prism\Providers\Groq\Maps\ToolChoiceMap;
 use EchoLabs\Prism\Providers\Groq\Maps\ToolMap;
-use EchoLabs\Prism\Providers\ProviderResponse;
 use EchoLabs\Prism\Text\Request;
+use EchoLabs\Prism\ValueObjects\ProviderResponse;
 use EchoLabs\Prism\ValueObjects\ToolCall;
 use EchoLabs\Prism\ValueObjects\Usage;
 use Illuminate\Http\Client\PendingRequest;
