+++ src/Providers/Mistral/Handlers/Text.php
@@ -9,8 +9,8 @@ use EchoLabs\Prism\Providers\Mistral\Maps\FinishReasonMap;
 use EchoLabs\Prism\Providers\Mistral\Maps\MessageMap;
 use EchoLabs\Prism\Providers\Mistral\Maps\ToolChoiceMap;
 use EchoLabs\Prism\Providers\Mistral\Maps\ToolMap;
-use EchoLabs\Prism\Providers\ProviderResponse;
 use EchoLabs\Prism\Text\Request;
+use EchoLabs\Prism\ValueObjects\ProviderResponse;
 use EchoLabs\Prism\ValueObjects\ToolCall;
 use EchoLabs\Prism\ValueObjects\Usage;
 use Illuminate\Http\Client\PendingRequest;
