+++ src/Text/Response.php
@@ -6,6 +6,7 @@ namespace EchoLabs\Prism\Text;
 
 use EchoLabs\Prism\Contracts\Message;
 use EchoLabs\Prism\Enums\FinishReason;
+use EchoLabs\Prism\ValueObjects\ResponseMeta;
 use EchoLabs\Prism\ValueObjects\ToolCall;
 use EchoLabs\Prism\ValueObjects\ToolResult;
 use EchoLabs\Prism\ValueObjects\Usage;
@@ -19,7 +20,6 @@ class Response
      * @param  ToolCall[]  $toolCalls
      * @param  ToolResult[]  $toolResults
      * @param  Collection<int, Message>  $messages
-     * @param  array{id: string, model: string}  $response
      */
     public function __construct(
         public readonly Collection $steps,
@@ -29,7 +29,7 @@ class Response
         public readonly array $toolCalls,
         public readonly array $toolResults,
         public readonly Usage $usage,
-        public readonly array $response,
+        public readonly ResponseMeta $responseMeta,
         public readonly Collection $messages,
     ) {}
 }
