+++ src/Text/Step.php
@@ -6,6 +6,7 @@ namespace EchoLabs\Prism\Text;
 
 use EchoLabs\Prism\Contracts\Message;
 use EchoLabs\Prism\Enums\FinishReason;
+use EchoLabs\Prism\ValueObjects\ResponseMeta;
 use EchoLabs\Prism\ValueObjects\ToolCall;
 use EchoLabs\Prism\ValueObjects\ToolResult;
 use EchoLabs\Prism\ValueObjects\Usage;
@@ -15,7 +16,6 @@ class Step
     /**
      * @param  ToolCall[]  $toolCalls
      * @param  ToolResult[]  $toolResults
-     * @param  array{id: string, model: string}  $response
      * @param  Message[]  $messages
      */
     public function __construct(
@@ -24,7 +24,7 @@ class Step
         public readonly array $toolCalls,
         public readonly array $toolResults,
         public readonly Usage $usage,
-        public readonly array $response,
+        public readonly ResponseMeta $responseMeta,
         public readonly array $messages,
     ) {}
 }
